+++ sphinxcontrib/needs.py
@@ -71,6 +71,7 @@ def setup(app):
     app.add_config_value('needs_id_required', False, 'html')
     app.add_config_value('needs_show_link_type', False, 'html')
     app.add_config_value('needs_show_link_title', False, 'html')
+    app.add_config_value('needs_file', False, 'html')
 
     app.add_config_value('needs_diagram_template',
                          DEFAULT_DIAGRAM_TEMPLATE,